Add validation checks
Source: https://github.com/msupply-foundation/open-msupply/commit/8f3253d38ccbdf70df200644876612e552bef002

+++ server/graphql/demographic/src/mutations/update_projection.rs
@@ -48,12 +48,12 @@ pub fn update_demographic_projection(
 #[derive(InputObject, Clone)]
 pub struct UpdateDemographicProjectionInput {
     pub id: String,
-    pub base_year: Option<i16>,
-    pub year_1: Option<f64>,
-    pub year_2: Option<f64>,
-    pub year_3: Option<f64>,
-    pub year_4: Option<f64>,
-    pub year_5: Option<f64>,
+    pub base_year: Option<i32>,
+    pub year_1: Option<i32>,
+    pub year_2: Option<i32>,
+    pub year_3: Option<i32>,
+    pub year_4: Option<i32>,
+    pub year_5: Option<i32>,
 }
 
 impl From<UpdateDemographicProjectionInput> for UpdateDemographicProjection {
